refactor(region): extract CivicsSyncService — #828 Step 5
Fifth bounded-context split. Extracts the civics-data ingestion
pipeline (declarative-source crawl → LLM extract → CivicsBlock +
glossary upserts) out of RegionSyncService.

region-sync.service.ts shrinks from 2,748 → 2,482 LOC (cumulative
3,929 → 2,482, −37%). CivicsSyncService is 367 LOC and absorbs four
methods plus a module-level helper:

  - syncCivics (the two-phase orchestration)
  - extractAndUpsertPage (was extractAndUpsertCivicsPage)
  - upsertGlossaryEntries (per-term glossary upserts)
  - toJsonField (Prisma JSONB null-handling helper, module-level)

Shared HTTP / HTML helpers (fetchUrlText, htmlToReadableText,
crawlCivicsUrls) stay in the orchestrator because bills sync uses them
too — passed in via a CivicsCrawlHelpers callback object per-call.
Bills extraction in Step 7 will lift them into a shared module as a
follow-up, removing the callback indirection.

The civics-ingest.service.spec.ts test file kept its canonicalizeUrl /
extractLinks / crawlCivicsUrls coverage (those helpers stay in the
orchestrator). The four syncCivics guard-path tests were deleted; the
equivalent guards live in CivicsSyncService now and are covered by
civics-sync.service.spec.ts in Step 9.

All 2,101 backend tests pass (was 2,105 — net −4 for the deleted
syncCivics guard tests; +0 added in this step because the new tests
land in Step 9).

Refs #828
